requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
PyYAML==6.0.2
imageio==2.36.1
imageio-ffmpeg==0.5.1
tqdm==4.67.1
matplotlib==3.9.2

